Press Done at the end of applying for a job so you can get back to searching! Removes Applied Jobs to get rid of the pesky old jobs cluttering your search. Removes hidden Jobs to get rid of the pesky irrelevant jobs cluttering your search. Automatically Dismisses Jobs Applied.
The above actions take place after you submit an application with Easy Apply Automator.

The next steps are to save jobs applied names and companies. Possibly using Redis.

The most important next step is to design how answers to questions are saved and their responses.
You can see some earlier attempts that will be integrated with RadioQuestions
Added getFormPageName which lets the program know which page is active when completing the application.